refactor: 초기 상태를 전역 상태의 이전 값으로 반영, 메모리 기능 추가
